+++ web/app/Domain/Repositories/ICardRepository.php
@@ -6,9 +6,20 @@ use App\Card;
 
 interface ICardRepository
 {
-
+    /**
+     * カード登録処理を行う関数
+     * 
+     * @param $data
+     * @return Card
+     */
     public function cardRegister($data): Card;
 
+
+    /**
+     * カード情報取得のための関数
+     * 
+     * @return array
+     */
     public function viewCard() :array; 
 
 }
